% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_set_asset.R
\name{assets_set_asset}
\alias{assets_set_asset}
\title{Set Asset}
\usage{
assets_set_asset(tok, assetname, refreshallclients = NULL)
}
\arguments{
\item{assetname}{Asset File to upload. Required}

\item{refreshallclients}{If needs refresh all clients to apply changes immediately. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Set an asset by name. Requires manage-assets permission.
}
