% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_unset_asset.R
\name{assets_unset_asset}
\alias{assets_unset_asset}
\title{Unset Asset}
\usage{
assets_unset_asset(tok, assetname, refreshallclients = NULL)
}
\arguments{
\item{assetname}{Asset naeme to unset. Required}

\item{refreshallclients}{If needs refresh all clients to apply changes immediately. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Unset an asset by name. Requires manage-assets permission.
}
