% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_login.R
\name{authentication_login}
\alias{authentication_login}
\title{Login}
\usage{
authentication_login(tok, user, password)
}
\arguments{
\item{user}{Your username or email Required}

\item{password}{Your password Required}

\item{token}{The token to connect to the app.}
}
\description{
Login with your username and password.
}
