% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_login_with_facebook.R
\name{authentication_login_with_facebook}
\alias{authentication_login_with_facebook}
\title{Login with Facebook}
\usage{
authentication_login_with_facebook(
  tok,
  servicename,
  accesstoken,
  secret,
  expiresin
)
}
\arguments{
\item{servicename}{The desired OAuth service name Required}

\item{accesstoken}{Access token provided by facebook oauth Required}

\item{secret}{The secret provided by facebook for app Required}

\item{expiresin}{Lifetime of token(in seconds) Required}

\item{token}{The token to connect to the app.}
}
\description{
Authenticate with Facebook.
}
