% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_login_with_google.R
\name{authentication_login_with_google}
\alias{authentication_login_with_google}
\title{Login with Google}
\usage{
authentication_login_with_google(
  tok,
  servicename,
  accesstoken,
  idtoken,
  expiresin,
  scope = NULL
)
}
\arguments{
\item{servicename}{The desired OAuth service name Required}

\item{accesstoken}{Access token provided by google oauth Required}

\item{idtoken}{Id token provided by google auth Required}

\item{expiresin}{Lifetime of token(in seconds) Required}

\item{scope}{Google scopes for API Optional}

\item{token}{The token to connect to the app.}
}
\description{
Authenticate with Google.
}
