% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_login_with_twitter.R
\name{authentication_login_with_twitter}
\alias{authentication_login_with_twitter}
\title{Login with Twitter}
\usage{
authentication_login_with_twitter(
  tok,
  servicename,
  accesstoken,
  accesstokensecret,
  appsecret,
  appid,
  expiresin
)
}
\arguments{
\item{servicename}{The desired OAuth service name Required}

\item{accesstoken}{Access token provided by twitter oauth Required}

\item{accesstokensecret}{Access token secret provided by twitter oauth Required}

\item{appsecret}{The secret provided by twitter for app Required}

\item{appid}{The id provided by twitter for app Required}

\item{expiresin}{Lifetime of token(in seconds) Required}

\item{token}{The token to connect to the app.}
}
\description{
Authenticate with Twitter.
}
