% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/autotranslate_auto_translate_get_supported_languages.R
\name{autotranslate_auto_translate_get_supported_languages}
\alias{autotranslate_auto_translate_get_supported_languages}
\title{Auto Translate Get Supported Languages}
\usage{
autotranslate_auto_translate_get_supported_languages(tok, targetlanguage)
}
\arguments{
\item{targetlanguage}{Language in which language names will be returned. No}

\item{token}{The token to connect to the app.}
}
\description{
Get the supported languages by the autotranslate.
}
