% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotranslate_auto_translate_save_settings.R
\name{autotranslate_auto_translate_save_settings}
\alias{autotranslate_auto_translate_save_settings}
\title{Auto Translate Save Settings}
\usage{
autotranslate_auto_translate_save_settings(
  tok,
  roomid,
  field,
  value,
  defaultlanguage = NULL
)
}
\arguments{
\item{roomid}{The room’s id to apply setting. Required}

\item{field}{The setting to apply to user’s subscription. Required}

\item{value}{Boolean if the setting is autoTranslate and a string (the language) if the setting is autoTranslateLanguage. Required}

\item{defaultlanguage}{The default language. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Save some settings about autotranslate.
}
