% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/autotranslate_auto_translate_translate_message.R
\name{autotranslate_auto_translate_translate_message}
\alias{autotranslate_auto_translate_translate_message}
\title{Auto Translate Translate Message}
\usage{
autotranslate_auto_translate_translate_message(
  tok,
  messageid,
  targetlanguage = NULL
)
}
\arguments{
\item{messageid}{The message’s id to be translated. Required}

\item{targetlanguage}{The target language that the message will be translated. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Translate the provided message.
}
