% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_list.R
\name{channel_list}
\alias{channel_list}
\title{Channel List}
\usage{
channel_list(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the channels on the server. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
