% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_list_joined.R
\name{channel_list_joined}
\alias{channel_list_joined}
\title{Channel List Joined}
\usage{
channel_list_joined(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the channels the calling user has joined. It supports the Offset, Count, and Sort Query Parameters along with just the Fields Query Parameter.
}
