% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_add_all.R
\name{channels_channel_add_all}
\alias{channels_channel_add_all}
\title{Channel Add All}
\usage{
channels_channel_add_all(tok, roomid, activeusersonly = NULL)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{activeusersonly}{Add active users only Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Adds all of the users of the Rocket.Chat server to the channel.
}
