% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_add_leader.R
\name{channels_channel_add_leader}
\alias{channels_channel_add_leader}
\title{Channel Add Leader}
\usage{
channels_channel_add_leader(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Gives the role of Leader for a user in the current channel.
}
