% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_add_moderator.R
\name{channels_channel_add_moderator}
\alias{channels_channel_add_moderator}
\title{Channel Add Moderator}
\usage{
channels_channel_add_moderator(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user id Required}

\item{token}{The token to connect to the app.}
}
\description{
Gives the role of moderator for a user in the current channel.
}
