% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_add_owner.R
\name{channels_channel_add_owner}
\alias{channels_channel_add_owner}
\title{Channel Add Owner}
\usage{
channels_channel_add_owner(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required(if no roomName}

\item{roomname}{The channel’s name Required(if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Gets the messages in public channels to an anonymous user, if the server has the setting Accounts_AllowAnonymousRead enabled.. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
