% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_archive.R
\name{channels_channel_archive}
\alias{channels_channel_archive}
\title{Channel Archive}
\usage{
channels_channel_archive(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Archives a channel.
}
