% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_close.R
\name{channels_channel_close}
\alias{channels_channel_close}
\title{Channel Close}
\usage{
channels_channel_close(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the channel from the user’s list of channels.
}
