% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_create.R
\name{channels_channel_create}
\alias{channels_channel_create}
\title{Channel Create}
\usage{
channels_channel_create(tok, name, members = NULL, readonly = NULL)
}
\arguments{
\item{name}{The name of the new channel Required}

\item{members}{The users to add to the channel when it is created. Optional  Default: []}

\item{readonly}{Set if the channel is read only or not. Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Creates a new public channel, optionally including specified users. The channel creator is always included.
}
