% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_delete.R
\name{channels_channel_delete}
\alias{channels_channel_delete}
\title{Channel Delete}
\usage{
channels_channel_delete(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required (if no roomName)}

\item{roomname}{The channel’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Remove a public channel.
}
