% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_files.R
\name{channels_channel_files}
\alias{channels_channel_files}
\title{Channel Files}
\usage{
channels_channel_files(tok, roomid_roomname)
}
\arguments{
\item{roomid_roomname}{The room id.  The room name. Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the files from a channel. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
