% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_history.R
\name{channels_channel_history}
\alias{channels_channel_history}
\title{Channel History}
\usage{
channels_channel_history(
  tok,
  roomid,
  latest = NULL,
  oldest = NULL,
  inclusive = NULL,
  offset = NULL,
  count = NULL,
  unreads = NULL
)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{latest}{The end of time range of messages to retrieve Optional  Default: now}

\item{oldest}{The start of the time range of messages to retrieve Optional  Default: n/a}

\item{inclusive}{Whether messages which land on latest and oldest should be included Optional  Default: false}

\item{offset}{The offset of the list messages to retrieve Optional  Default: 0}

\item{count}{The amount of messages to retrieve Optional  Default: 20}

\item{unreads}{Whether the amount of unreads should be included. Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the messages from a channel.
}
