% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_info.R
\name{channels_channel_info}
\alias{channels_channel_info}
\title{Channel Info}
\usage{
channels_channel_info(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required (if no roomName)}

\item{roomname}{The channel’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the information about the channel.
}
