% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_invite.R
\name{channels_channel_invite}
\alias{channels_channel_invite}
\title{Channel Invite}
\usage{
channels_channel_invite(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user id of who to invite Required}

\item{token}{The token to connect to the app.}
}
\description{
Adds a user to the channel.
}
