% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_join.R
\name{channels_channel_join}
\alias{channels_channel_join}
\title{Channel Join}
\usage{
channels_channel_join(tok, roomid, joincode)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{joincode}{The join code of the room Required}

\item{token}{The token to connect to the app.}
}
\description{
Joins yourself to the channel.
}
