% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_kick.R
\name{channels_channel_kick}
\alias{channels_channel_kick}
\title{Channel Kick}
\usage{
channels_channel_kick(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user id of who to kick Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes a user from the channel.
}
