% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_leave.R
\name{channels_channel_leave}
\alias{channels_channel_leave}
\title{Channel Leave}
\usage{
channels_channel_leave(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Causes the callee to be removed from the channel.
}
