% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_moderators_list.R
\name{channels_channel_moderators_list}
\alias{channels_channel_moderators_list}
\title{Channel Moderators List}
\usage{
channels_channel_moderators_list(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required (if no roomName)}

\item{roomname}{The channel’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all channel moderators.
}
