% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_online.R
\name{channels_channel_online}
\alias{channels_channel_online}
\title{Channel Online}
\usage{
channels_channel_online(tok, query = NULL)
}
\arguments{
\item{query}{See Query Parameter Optional}

\item{token}{The token to connect to the app.}
}
\description{
Lists all online users of a channel if the channel’s id is provided, otherwise it gets all online users of all channels. It supports the Query Parameters only.
}
