% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_open.R
\name{channels_channel_open}
\alias{channels_channel_open}
\title{Channel Open}
\usage{
channels_channel_open(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Adds the channel back to the user’s list of channels.
}
