% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_remove_moderator.R
\name{channels_channel_remove_moderator}
\alias{channels_channel_remove_moderator}
\title{Channel Remove Moderator}
\usage{
channels_channel_remove_moderator(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the role of moderator from a user in the current channel.
}
