% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_remove_owner.R
\name{channels_channel_remove_owner}
\alias{channels_channel_remove_owner}
\title{Channel Remove Owner}
\usage{
channels_channel_remove_owner(tok, roomid, userid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the role of owner from a user in the current channel.
}
