% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_rename.R
\name{channels_channel_rename}
\alias{channels_channel_rename}
\title{Channel Rename}
\usage{
channels_channel_rename(tok, roomid, name)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{name}{The new name the channel will have, it cannot be the same as another or the existing name Required}

\item{token}{The token to connect to the app.}
}
\description{
Changes the name of the channel.
}
