% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_roles.R
\name{channels_channel_roles}
\alias{channels_channel_roles}
\title{Channel Roles}
\usage{
channels_channel_roles(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required(if no roomName)}

\item{roomname}{The channel’s name Required(if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all user’s roles in the channel.
}
