% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_announcement.R
\name{channels_channel_set_announcement}
\alias{channels_channel_set_announcement}
\title{Channel Set Announcement}
\usage{
channels_channel_set_announcement(tok, roomid, announcement)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{announcement}{The announcement to set for the channel. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the announcement for the channel.
}
