% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_custom_fields.R
\name{channels_channel_set_custom_fields}
\alias{channels_channel_set_custom_fields}
\title{Channel Set Custom Fields}
\usage{
channels_channel_set_custom_fields(tok, roomid, roomname, customfields)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{roomname}{The channel’s name Required}

\item{customfields}{The custom fields to set for the channel. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the custom fields for the channel.
}
