% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_default.R
\name{channels_channel_set_default}
\alias{channels_channel_set_default}
\title{Channel Set Default}
\usage{
channels_channel_set_default(tok, roomid, default)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{default}{Boolean of whether the room is default or not. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets whether the channel is a default channel or not.
}
