% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_description.R
\name{channels_channel_set_description}
\alias{channels_channel_set_description}
\title{Channel Set Description}
\usage{
channels_channel_set_description(tok, roomid, description)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{description}{The description to set for the channel. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the description for the channel.
}
