% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_join_code.R
\name{channels_channel_set_join_code}
\alias{channels_channel_set_join_code}
\title{Channel Set Join Code}
\usage{
channels_channel_set_join_code(tok, roomid, joincode)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{joincode}{The join code which this channel will require to be joined. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the code required to join the channel.
}
