% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_purpose.R
\name{channels_channel_set_purpose}
\alias{channels_channel_set_purpose}
\title{Channel Set Purpose}
\usage{
channels_channel_set_purpose(tok, roomid, purpose)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{purpose}{The description to set for the channel. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the description for the channel (the same as channels.setDescription, obsolete).
}
