% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_read_only.R
\name{channels_channel_set_read_only}
\alias{channels_channel_set_read_only}
\title{Channel Set Read Only}
\usage{
channels_channel_set_read_only(tok, roomid, readonly)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{readonly}{Boolean of whether the room is read only or not. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets whether the channel is read only or not.
}
