% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_topic.R
\name{channels_channel_set_topic}
\alias{channels_channel_set_topic}
\title{Channel Set Topic}
\usage{
channels_channel_set_topic(tok, roomid, topic)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{topic}{The channel’s topic to set. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the topic for the channel.
}
