% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_set_type.R
\name{channels_channel_set_type}
\alias{channels_channel_set_type}
\title{Channel Set Type}
\usage{
channels_channel_set_type(tok, roomid, type)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{type}{The type of room this channel should be, either c or p. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the type of room this channel should be.
}
