% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_channel_unarchive.R
\name{channels_channel_unarchive}
\alias{channels_channel_unarchive}
\title{Channel Unarchive}
\usage{
channels_channel_unarchive(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Unarchives a channel.
}
