% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_get_all_the_mentions_of_a_channel.R
\name{channels_get_all_the_mentions_of_a_channel}
\alias{channels_get_all_the_mentions_of_a_channel}
\title{Get all the mentions of a channel}
\usage{
channels_get_all_the_mentions_of_a_channel(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
It supports the Offset, Count, and Sort Query Parameters.
}
