% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels_get_integrations.R
\name{channels_get_integrations}
\alias{channels_get_integrations}
\title{Get Integrations}
\usage{
channels_get_integrations(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the integrations which the channel has, requires at least one integration permission: manage-incoming-integrations, manage-own-incoming-integrations, manage-outgoing-integrations or manage-own-outgoing-integrations. It will return the integrations based on the user permission. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
