% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_chat_message_delete.R
\name{chat_chat_message_delete}
\alias{chat_chat_message_delete}
\title{Chat Message Delete}
\usage{
chat_chat_message_delete(tok, roomid, msgid, asuser = NULL)
}
\arguments{
\item{roomid}{The room id of where the message is to delete. Required}

\item{msgid}{The message id to delete. Required}

\item{asuser}{Whether the message should be deleted as the user who sent it. Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Chat Message Delete
}
