% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_chat_message_reactions.R
\name{chat_chat_message_reactions}
\alias{chat_chat_message_reactions}
\title{Chat Message Reactions}
\usage{
chat_chat_message_reactions(tok, emoji, messageid, shouldreact = NULL)
}
\arguments{
\item{emoji}{The emoji to react with. Required}

\item{messageid}{The message id to react to. Required}

\item{shouldreact}{Parameter to specify if the reaction should be added or removed. Optional: true or false}

\item{token}{The token to connect to the app.}
}
\description{
Toggles the authenticated user’s reaction to the provided message.
}
