% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_chat_message_update.R
\name{chat_chat_message_update}
\alias{chat_chat_message_update}
\title{Chat Message Update}
\usage{
chat_chat_message_update(tok, roomid, msgid, text)
}
\arguments{
\item{roomid}{The room id of where the message is. Required}

\item{msgid}{The message id to update. Required}

\item{text}{Updated text for the message. Required}

\item{token}{The token to connect to the app.}
}
\description{
Chat Message Update
}
