% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_chat_report_message.R
\name{chat_chat_report_message}
\alias{chat_chat_report_message}
\title{Chat Report Message}
\usage{
chat_chat_report_message(tok, messageid, description)
}
\arguments{
\item{messageid}{The message id to report. Required}

\item{description}{The description of report. Required}

\item{token}{The token to connect to the app.}
}
\description{
Reports a message.
}
