% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_follows_a_chat_message.R
\name{chat_follows_a_chat_message}
\alias{chat_follows_a_chat_message}
\title{Follows a Chat Message}
\usage{
chat_follows_a_chat_message(tok, mid)
}
\arguments{
\item{mid}{The message id to follow. Required}

\item{token}{The token to connect to the app.}
}
\description{
Follows a chat message to the message’s channel.
}
