% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_channel_threads.R
\name{chat_get_channel_threads}
\alias{chat_get_channel_threads}
\title{Get channel threads}
\usage{
chat_get_channel_threads(tok, rid)
}
\arguments{
\item{rid}{The id of the channel. Required}

\item{token}{The token to connect to the app.}
}
\description{
Get channel threads. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
