% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_deleted_messages.R
\name{chat_get_deleted_messages}
\alias{chat_get_deleted_messages}
\title{Get Deleted Messages}
\usage{
chat_get_deleted_messages(tok, roomid, since)
}
\arguments{
\item{roomid}{The room id Required}

\item{since}{The date (ISODate) to search deleted messages Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve message read receipts. It supports ONLY the Offset and Count Query Parameters
}
