% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_discussions_of_a_room.R
\name{chat_get_discussions_of_a_room}
\alias{chat_get_discussions_of_a_room}
\title{Get Discussions of A Room}
\usage{
chat_get_discussions_of_a_room(tok, roomid)
}
\arguments{
\item{roomid}{The room id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve the discussions of a room. It supports the Offset, Count, and Sort Query Parameters.
}
