% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_message_read_receipts.R
\name{chat_get_message_read_receipts}
\alias{chat_get_message_read_receipts}
\title{Get Message Read Receipts}
\usage{
chat_get_message_read_receipts(tok, messageid)
}
\arguments{
\item{messageid}{The message id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve message read receipts.
}
