% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_pinned_messages.R
\name{chat_get_pinned_messages}
\alias{chat_get_pinned_messages}
\title{Get Pinned Messages}
\usage{
chat_get_pinned_messages(tok, roomid, offset = NULL, count = NULL)
}
\arguments{
\item{roomid}{The room id Required}

\item{offset}{See Offset Parameter Optional}

\item{count}{See Count Parameter Optional}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve pinned messages from a room. It supports ONLY the Offset and Count Query Parameters
}
