% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_snippeted_message_by_id.R
\name{chat_get_snippeted_message_by_id}
\alias{chat_get_snippeted_message_by_id}
\title{Get Snippeted Message By Id}
\usage{
chat_get_snippeted_message_by_id(tok, messageid)
}
\arguments{
\item{messageid}{The message id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve snippeted message by Id.
}
