% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_snippeted_messages.R
\name{chat_get_snippeted_messages}
\alias{chat_get_snippeted_messages}
\title{Get Snippeted Messages}
\usage{
chat_get_snippeted_messages(tok, roomid)
}
\arguments{
\item{roomid}{The room id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieve snippeted messages. It supports the Offset, Count, and Sort Query Parameters.
}
