% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_get_thread_messages.R
\name{chat_get_thread_messages}
\alias{chat_get_thread_messages}
\title{Get thread messages}
\usage{
chat_get_thread_messages(tok, tmid)
}
\arguments{
\item{tmid}{The id of the the thread message. Required}

\item{token}{The token to connect to the app.}
}
\description{
Get thread messages. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
