% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_ignore_user.R
\name{chat_ignore_user}
\alias{chat_ignore_user}
\title{Ignore user}
\usage{
chat_ignore_user(tok, rid, userid, ignore = NULL)
}
\arguments{
\item{rid}{The room ID. Required}

\item{userid}{The User ID. Required}

\item{ignore}{If the user will be ignored or not, default is true. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Ignores an user in a chat. If you pass ignore as false, the user will be unignored.
}
