% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_pins_a_chat_message.R
\name{chat_pins_a_chat_message}
\alias{chat_pins_a_chat_message}
\title{Pins a Chat Message}
\usage{
chat_pins_a_chat_message(tok, messageid)
}
\arguments{
\item{messageid}{The message id to pin. Required}

\item{token}{The token to connect to the app.}
}
\description{
Pins a chat message to the message’s channel.
}
