% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_post_a_chat_message.R
\name{chat_post_a_chat_message}
\alias{chat_post_a_chat_message}
\title{Post a chat message}
\usage{
chat_post_a_chat_message(
  tok,
  roomid_channel,
  text = NULL,
  alias = NULL,
  emoji = NULL,
  avatar = NULL,
  attachments = NULL
)
}
\arguments{
\item{roomid_channel}{The room id of where the message is to be sent.  The channel name with the prefix in front of it. Required}

\item{text}{The text of the message to send, is optional because of attachments. Optional}

\item{alias}{This will cause the message’s name to appear as the given alias, but your username will still display. Optional}

\item{emoji}{If provided, this will make the avatar on this message be an emoji. Emoji Cheetsheet Optional}

\item{avatar}{If provided, this will make the avatar use the provided image url. Optional}

\item{attachments}{See the below section, Attachments Detail, for details. Optional}

\item{token}{The token to connect to the app.}
}
\description{
You only can send alias and avatar properties if your user has the bot role. We implement this rule to avoid users to impersonate other users.
}
