% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_search_for_messages_in_a_channel.R
\name{chat_search_for_messages_in_a_channel}
\alias{chat_search_for_messages_in_a_channel}
\title{Search for messages in a channel}
\usage{
chat_search_for_messages_in_a_channel(tok, roomid, searchtext, count = NULL)
}
\arguments{
\item{roomid}{The id of the channel. Required}

\item{searchtext}{The text message to search in messages. Required}

\item{count}{The limit of result. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Search for messages in a channel by id and text message.
}
