% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_send_new_chat_message.R
\name{chat_send_new_chat_message}
\alias{chat_send_new_chat_message}
\title{Send new chat message}
\usage{
chat_send_new_chat_message(
  tok,
  message_id,
  message_rid,
  message_tmid = NULL,
  message_msg = NULL,
  message_alias = NULL,
  message_emoji = NULL,
  message_avatar = NULL,
  message_attachments = NULL
)
}
\arguments{
\item{message_id}{The _id of message. Required}

\item{message_rid}{The room id of where the message is to be sent. Required}

\item{message_tmid}{The message’s id to create a thread. Optional}

\item{message_msg}{The text of the message to send, is optional because of attachments. Optional}

\item{message_alias}{This will cause the message’s name to appear as the given alias, but your username will still display. Optional}

\item{message_emoji}{If provided, this will make the avatar on this message be an emoji. Emoji Cheetsheet Optional}

\item{message_avatar}{If provided, this will make the avatar use the provided image url. Optional}

\item{message_attachments}{See the below section, Attachments Detail, for details. Optional}

\item{token}{The token to connect to the app.}
}
\description{
The difference between chat.postMessage and chat.sendMessage is that chat.sendMessage allows. for passing a value for _id and the other one doesn’t. Also, chat.sendMessage only sends it to. one channel whereas the other one allows for sending to more than one channel at a time.
}
