% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_star_a_chat_message.R
\name{chat_star_a_chat_message}
\alias{chat_star_a_chat_message}
\title{Star a Chat Message}
\usage{
chat_star_a_chat_message(tok, messageid)
}
\arguments{
\item{messageid}{The message id to star. Required}

\item{token}{The token to connect to the app.}
}
\description{
Stars a chat message for the authenticated user.
}
