% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_synced_thread_list.R
\name{chat_synced_thread_list}
\alias{chat_synced_thread_list}
\title{Synced thread list}
\usage{
chat_synced_thread_list(tok, rid, updatedsince)
}
\arguments{
\item{rid}{Room’s ID Required}

\item{updatedsince}{Date as ISO string Required}

\item{token}{The token to connect to the app.}
}
\description{
List all custom emojis available. It supports the Query, Fields and sort.
}
