% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_synced_thread_s_messages.R
\name{chat_synced_thread_s_messages}
\alias{chat_synced_thread_s_messages}
\title{Synced thread’s messages}
\usage{
chat_synced_thread_s_messages(tok, tmid, updatedsince)
}
\arguments{
\item{tmid}{Thread’s message ID Required}

\item{updatedsince}{Date as ISO string Required}

\item{token}{The token to connect to the app.}
}
\description{
List all custom emojis available. It supports the Query, Fields and sort.
}
