% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_unfollow_a_chat_message.R
\name{chat_unfollow_a_chat_message}
\alias{chat_unfollow_a_chat_message}
\title{Unfollow a Chat Message}
\usage{
chat_unfollow_a_chat_message(tok, mid)
}
\arguments{
\item{mid}{The message id to unfollow. Required}

\item{token}{The token to connect to the app.}
}
\description{
Unfollow a chat message to the message’s channel.
}
