% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_unpins_a_chat_message.R
\name{chat_unpins_a_chat_message}
\alias{chat_unpins_a_chat_message}
\title{Unpins a Chat Message}
\usage{
chat_unpins_a_chat_message(tok, messageid)
}
\arguments{
\item{messageid}{The message id to unpin. Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the pinned status of the provided chat message.
}
