% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_unstar_a_chat_message.R
\name{chat_unstar_a_chat_message}
\alias{chat_unstar_a_chat_message}
\title{Unstar a Chat Message}
\usage{
chat_unstar_a_chat_message(tok, messageid)
}
\arguments{
\item{messageid}{The message id to unpin. Required}

\item{token}{The token to connect to the app.}
}
\description{
Unstars the chat message for the authenticated user.
}
