% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create}
\usage{
create(tok, email, name, password, username, active = NULL,
  roles = NULL, joindefaultchannels = NULL,
  requirepasswordchange = NULL, sendwelcomeemail = NULL,
  verified = NULL, customfields = NULL)
}
\arguments{
\item{email}{The email address for the user. Required}

\item{name}{The display name of the user. Required}

\item{password}{The password for the user. Required}

\item{username}{The username for the user. Required}

\item{active}{Whether the user is active, which determines if they can login or not. Optional  Default: true}

\item{roles}{The roles the user has assigned to them on creation. Optional  Default: ['user']}

\item{joindefaultchannels}{Whether the user should join the default channels when created. Optional  Default: true}

\item{requirepasswordchange}{Should the user be required to change their password when they login? Optional  Default: false}

\item{sendwelcomeemail}{Should the user get a welcome email? Optional  Default: false}

\item{verified}{Should the user’s email address be verified when created? Optional  Default: false}

\item{customfields}{Any custom fields the user should have on their account. Optional  Default: undefined}

\item{token}{The token to connect to the app.}
}
\description{
Create a new user. Requires create-user permission.
}
