% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_sounds_custom_sounds_list.R
\name{custom_sounds_list}
\alias{custom_sounds_list}
\title{Custom Sounds List}
\usage{
custom_sounds_list(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Retrieves all custom sounds. It supports the Offset, Count, and Sort Query Parameters along with Query Parameters.
}
