% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_user_status_custom_user_status_list.R
\name{custom_user_status_list}
\alias{custom_user_status_list}
\title{Custom User Status List}
\usage{
custom_user_status_list(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all available custom user’s status.
}
