% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_dealing_with_livechat_custom_fields.R
\name{dealing_with_livechat_custom_fields}
\alias{dealing_with_livechat_custom_fields}
\title{Dealing with Livechat Custom Fields}
\usage{
dealing_with_livechat_custom_fields(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Get a list of Livechat custom fields. It supports the Offset, Count, and Sort Query Parameters.
}
