% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete}
\usage{
delete(tok, userid_or_username)
}
\arguments{
\item{userid_or_username}{The id or username of the user. Required}

\item{token}{The token to connect to the app.}
}
\description{
Deletes an existing user. Requires delete-user permission.
}
