% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_custom_emoji_custom_create.R
\name{emoji_custom_emoji_custom_create}
\alias{emoji_custom_emoji_custom_create}
\title{Emoji Custom create}
\usage{
emoji_custom_emoji_custom_create(tok, emoji, name, aliases)
}
\arguments{
\item{emoji}{The image file to use as the new custom emoji, as form data. Required}

\item{name}{The name of the new custom emoji. Required}

\item{aliases}{The alias of the new custom emoji. Required}

\item{token}{The token to connect to the app.}
}
\description{
Create new custom emoji.
}
