% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_custom_emoji_custom_delete.R
\name{emoji_custom_emoji_custom_delete}
\alias{emoji_custom_emoji_custom_delete}
\title{Emoji Custom delete}
\usage{
emoji_custom_emoji_custom_delete(tok, emojiid)
}
\arguments{
\item{emojiid}{The custom emoji id to be deleted. Required}

\item{token}{The token to connect to the app.}
}
\description{
Emoji Custom delete
}
