% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_custom_emoji_custom_list.R
\name{emoji_custom_emoji_custom_list}
\alias{emoji_custom_emoji_custom_list}
\title{Emoji custom list}
\usage{
emoji_custom_emoji_custom_list(tok, updatedsince = NULL)
}
\arguments{
\item{updatedsince}{Date as ISO string Optional}

\item{token}{The token to connect to the app.}
}
\description{
List all custom emojis available. It supports the Query Parameters only.
}
