% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_custom_emoji_custom_update.R
\name{emoji_custom_emoji_custom_update}
\alias{emoji_custom_emoji_custom_update}
\title{Emoji Custom update}
\usage{
emoji_custom_emoji_custom_update(tok, emoji = NULL, x_id, name, aliases = NULL)
}
\arguments{
\item{emoji}{The image file to use as custom emoji, as form data. Optional}

\item{x_id}{The id of the custom emoji. Required}

\item{name}{The name of the custom emoji. Required}

\item{aliases}{The alias of the custom emoji. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Update an existent custom emoji.
}
