% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_personal_access_token.R
\name{generate_personal_access_token}
\alias{generate_personal_access_token}
\title{Generate Personal Access Token}
\usage{
generate_personal_access_token(tok, tokenname)
}
\arguments{
\item{tokenname}{The name of the token. Required}

\item{token}{The token to connect to the app.}
}
\description{
Generate Personal Access Token. Requires create-personal-access-tokens permission.
}
