% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_get_personal_access_tokens.R
\name{get_personal_access_tokens}
\alias{get_personal_access_tokens}
\title{Get Personal Access Tokens}
\usage{
get_personal_access_tokens(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Gets the user’s personal access tokens. Requires create-personal-access-tokens permission.
}
