% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preferences.R
\name{get_preferences}
\alias{get_preferences}
\title{Get Preferences}
\usage{
get_preferences(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Gets all preferences of the user.
}
