% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_get_username_suggestion.R
\name{get_username_suggestion}
\alias{get_username_suggestion}
\title{Get Username Suggestion}
\usage{
get_username_suggestion(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Gets a suggestion a new username to user.
}
