% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_list.R
\name{group_list}
\alias{group_list}
\title{Group List}
\usage{
group_list(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the private groups the calling user has joined. It supports the Offset, Count, and Sort Query Parameters.
}
