% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_list_all.R
\name{group_list_all}
\alias{group_list_all}
\title{Group List All}
\usage{
group_list_all(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the private groups of any users. The calling user requires to have ‘view-room-administration’ right. It supports the Offset, Count, and Sort Query Parameters along with just the Fields Query Parameters.
}
