% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_add_all.R
\name{groups_group_add_all}
\alias{groups_group_add_all}
\title{Group Add All}
\usage{
groups_group_add_all(tok, roomid, activeusersonly = NULL)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{activeusersonly}{Add active users only Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Adds all of the users of the Rocket.Chat server to the group.
}
