% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_add_leader.R
\name{groups_group_add_leader}
\alias{groups_group_add_leader}
\title{Group Add Leader}
\usage{
groups_group_add_leader(tok, roomid, userid)
}
\arguments{
\item{roomid}{The groups’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Gives the role of Leader for a user in the current group.
}
