% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_add_moderator.R
\name{groups_group_add_moderator}
\alias{groups_group_add_moderator}
\title{Group Add Moderator}
\usage{
groups_group_add_moderator(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user id Required}

\item{token}{The token to connect to the app.}
}
\description{
Gives the role of moderator for a user in the current group.
}
