% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_add_owner.R
\name{groups_group_add_owner}
\alias{groups_group_add_owner}
\title{Group Add Owner}
\usage{
groups_group_add_owner(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user id Required}

\item{token}{The token to connect to the app.}
}
\description{
Gives the role of owner for a user in the current group.
}
