% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_archive.R
\name{groups_group_archive}
\alias{groups_group_archive}
\title{Group Archive}
\usage{
groups_group_archive(tok, roomid)
}
\arguments{
\item{roomid}{The private groups’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Archives a private group, only if you’re part of the group.
}
