% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_close.R
\name{groups_group_close}
\alias{groups_group_close}
\title{Group Close}
\usage{
groups_group_close(tok, roomid)
}
\arguments{
\item{roomid}{The groups’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the private group from the user’s list of groups, only if you’re part of the group.
}
