% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_counters.R
\name{groups_group_counters}
\alias{groups_group_counters}
\title{Group Counters}
\usage{
groups_group_counters(tok, roomid, roomname, userid = NULL)
}
\arguments{
\item{roomid}{The group’s id Required}

\item{roomname}{The group’s name Required}

\item{userid}{Counters for provided user id (need to have a view-room-administration right for calling user) Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets group counters.
}
