% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_create.R
\name{groups_group_create}
\alias{groups_group_create}
\title{Group Create}
\usage{
groups_group_create(tok, name, members = NULL, readonly = NULL)
}
\arguments{
\item{name}{The name of the new private group Required}

\item{members}{The users to add to the group when it is created. Optional  Default: []}

\item{readonly}{Set if the group is read only or not. Optional  Default: false}

\item{token}{The token to connect to the app.}
}
\description{
Creates a new private group, optionally including specified users. The group creator is always included.
}
