% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_delete.R
\name{groups_group_delete}
\alias{groups_group_delete}
\title{Group Delete}
\usage{
groups_group_delete(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The channel’s id Required (if no roomName)}

\item{roomname}{The channel’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Remove a private channel.
}
