% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_info.R
\name{groups_group_info}
\alias{groups_group_info}
\title{Group Info}
\usage{
groups_group_info(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The private group’s id Required (if no roomName)}

\item{roomname}{The private group’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the information about the private group, only if you’re part of the group.
}
