% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_invite.R
\name{groups_group_invite}
\alias{groups_group_invite}
\title{Group Invite}
\usage{
groups_group_invite(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user id of who to invite Required}

\item{token}{The token to connect to the app.}
}
\description{
Adds a user to the private group.
}
