% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_kick.R
\name{groups_group_kick}
\alias{groups_group_kick}
\title{Group Kick}
\usage{
groups_group_kick(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user id of who to invite Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes a user from the private group.
}
