% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_leave.R
\name{groups_group_leave}
\alias{groups_group_leave}
\title{Group Leave}
\usage{
groups_group_leave(tok, roomid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Causes the callee to be removed from the private group, if they’re part of it and are not the last owner.
}
