% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_members.R
\name{groups_group_members}
\alias{groups_group_members}
\title{Group Members}
\usage{
groups_group_members(tok, roomid_roomname)
}
\arguments{
\item{roomid_roomname}{The room id.  The room name. Required}

\item{token}{The token to connect to the app.}
}
\description{
Lists the users of participants of a private group. It supports the Offset, Count, and Sort Query Parameters.
}
