% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_messages.R
\name{groups_group_messages}
\alias{groups_group_messages}
\title{Group messages}
\usage{
groups_group_messages(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The private group’s id Required(if no roomName}

\item{roomname}{The private group’s name Required(if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all of the specific group messages on the server. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
