% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_moderators_list.R
\name{groups_group_moderators_list}
\alias{groups_group_moderators_list}
\title{Group Moderators List}
\usage{
groups_group_moderators_list(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The group’s id Required (if no roomName)}

\item{roomname}{The group’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all group moderators.
}
