% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_open.R
\name{groups_group_open}
\alias{groups_group_open}
\title{Group Open}
\usage{
groups_group_open(tok, roomid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Adds the private group back to the user’s list of private groups.
}
