% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_remove_moderator.R
\name{groups_group_remove_moderator}
\alias{groups_group_remove_moderator}
\title{Group Remove Moderator}
\usage{
groups_group_remove_moderator(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the role of moderator from a user in the current group.
}
