% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_remove_owner.R
\name{groups_group_remove_owner}
\alias{groups_group_remove_owner}
\title{Group Remove Owner}
\usage{
groups_group_remove_owner(tok, roomid, userid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{userid}{The user’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the role of owner from a user in the current Group.
}
