% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_rename.R
\name{groups_group_rename}
\alias{groups_group_rename}
\title{Group Rename}
\usage{
groups_group_rename(tok, roomid, name)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{name}{The new name the private group will have, it can not be the same as another or the existing name Required}

\item{token}{The token to connect to the app.}
}
\description{
Changes the name of the private group.
}
