% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_announcement.R
\name{groups_group_set_announcement}
\alias{groups_group_set_announcement}
\title{Group Set Announcement}
\usage{
groups_group_set_announcement(tok, roomid, announcement)
}
\arguments{
\item{roomid}{The group’s id Required}

\item{announcement}{The announcement to set for the group. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the announcement for the group.
}
