% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_custom_fields.R
\name{groups_group_set_custom_fields}
\alias{groups_group_set_custom_fields}
\title{Group Set Custom Fields}
\usage{
groups_group_set_custom_fields(tok, roomid, roomname, customfields)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{roomname}{The private group’s name Required}

\item{customfields}{The custom fields to set for the private group. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the custom fields for the private group.
}
