% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_description.R
\name{groups_group_set_description}
\alias{groups_group_set_description}
\title{Group Set Description}
\usage{
groups_group_set_description(tok, roomid, description)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{description}{The description to set for the private group. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the description for the private group.
}
