% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_purpose.R
\name{groups_group_set_purpose}
\alias{groups_group_set_purpose}
\title{Group Set Purpose}
\usage{
groups_group_set_purpose(tok, roomid, purpose)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{purpose}{The description to set for the private group. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the description for the private group (the same as groups.setDescription, obsolete).
}
