% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_read_only.R
\name{groups_group_set_read_only}
\alias{groups_group_set_read_only}
\title{Group Set Read Only}
\usage{
groups_group_set_read_only(tok, roomid, readonly)
}
\arguments{
\item{roomid}{The group’s id Required}

\item{readonly}{Boolean of whether the group is read only or not. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets whether the group is read only or not.
}
