% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_topic.R
\name{groups_group_set_topic}
\alias{groups_group_set_topic}
\title{Group Set Topic}
\usage{
groups_group_set_topic(tok, roomid, topic)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{topic}{The private group’s topic to set. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the topic for the private group.
}
