% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_set_type.R
\name{groups_group_set_type}
\alias{groups_group_set_type}
\title{Group Set Type}
\usage{
groups_group_set_type(tok, roomid, type)
}
\arguments{
\item{roomid}{The group’s id Required}

\item{type}{The type of room this group should be, either c or p. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the type of room this group should be.
}
