% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_group_unarchive.R
\name{groups_group_unarchive}
\alias{groups_group_unarchive}
\title{Group Unarchive}
\usage{
groups_group_unarchive(tok, roomid)
}
\arguments{
\item{roomid}{The private group’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Unarchives a private group.
}
