% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_groups_files.R
\name{groups_groups_files}
\alias{groups_groups_files}
\title{Groups Files}
\usage{
groups_groups_files(tok, roomid_roomname)
}
\arguments{
\item{roomid_roomname}{The room id.  The room name. Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the files from a private group. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
