% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_groups_roles.R
\name{groups_groups_roles}
\alias{groups_groups_roles}
\title{Groups Roles}
\usage{
groups_groups_roles(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The group’s id Required(if no roomName)}

\item{roomname}{The group’s name Required(if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all user’s roles in the private group.
}
