% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_counters_of_direct_messages.R
\name{im_counters_of_direct_messages}
\alias{im_counters_of_direct_messages}
\title{Counters of Direct Messages}
\usage{
im_counters_of_direct_messages(tok, roomid, username = NULL)
}
\arguments{
\item{roomid}{The id of Direct Messages’ Room Required}

\item{username}{Counters for provided username (need to have a view-room-administration right for calling user) Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets counters of direct messages.
}
