% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_close.R
\name{im_im_close}
\alias{im_im_close}
\title{IM Close}
\usage{
im_im_close(tok, roomid)
}
\arguments{
\item{roomid}{The direct message id Required}

\item{token}{The token to connect to the app.}
}
\description{
Removes the direct message from the user’s list of direct messages.
}
