% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_create.R
\name{im_im_create}
\alias{im_im_create}
\title{IM Create}
\usage{
im_im_create(tok, username)
}
\arguments{
\item{username}{The username to open a direct message session Required}

\item{token}{The token to connect to the app.}
}
\description{
Create a direct message session with another user.
}
