% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_files.R
\name{im_im_files}
\alias{im_im_files}
\title{Im Files}
\usage{
im_im_files(tok, roomid_username)
}
\arguments{
\item{roomid_username}{The room id.  The user name. Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the files from a direct message. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
