% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_members.R
\name{im_im_members}
\alias{im_im_members}
\title{IM Members}
\usage{
im_im_members(tok, roomid_username)
}
\arguments{
\item{roomid_username}{The room id.  The user name. Required}

\item{token}{The token to connect to the app.}
}
\description{
Lists the users of participants of a direct message. It supports the Offset, Count, and Sort Query Parameters.
}
