% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_messages.R
\name{im_im_messages}
\alias{im_im_messages}
\title{Im messages}
\usage{
im_im_messages(tok, roomid, username)
}
\arguments{
\item{roomid}{The channel’s id Required(if no username)}

\item{username}{The username of DM Required(if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Lists all of the specific direct message on the server. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
