% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_messages_others.R
\name{im_im_messages_others}
\alias{im_im_messages_others}
\title{IM Messages Others}
\usage{
im_im_messages_others(tok, roomid)
}
\arguments{
\item{roomid}{The direct message id Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the messages from any direct message in the server. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameter.
}
