% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_open.R
\name{im_im_open}
\alias{im_im_open}
\title{IM Open}
\usage{
im_im_open(tok, roomid)
}
\arguments{
\item{roomid}{The direct message id Required}

\item{token}{The token to connect to the app.}
}
\description{
Adds the direct message back to the user’s list of direct messages.
}
