% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_set_topic.R
\name{im_im_set_topic}
\alias{im_im_set_topic}
\title{IM Set Topic}
\usage{
im_im_set_topic(tok, roomid, topic)
}
\arguments{
\item{roomid}{The direct message id Required}

\item{topic}{The direct message topic to set. Required}

\item{token}{The token to connect to the app.}
}
\description{
Sets the topic for the direct message.
}
