% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_list.R
\name{im_list}
\alias{im_list}
\title{IM List}
\usage{
im_list(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the direct messages the calling user has joined. It supports the Offset, Count, and Sort Query Parameters.
}
