% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_im_list_everyone.R
\name{im_list_everyone}
\alias{im_list_everyone}
\title{IM List Everyone}
\usage{
im_list_everyone(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all of the direct messages in the server, requires the permission view-room-administration permission. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameters.
}
