% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_info.R
\name{info}
\alias{info}
\title{Info}
\usage{
info(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
A simple method, requires no authentication, that returns information about the server including version information.
}
