% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_integrations_settings.R
\name{integrations_settings}
\alias{integrations_settings}
\title{Integrations Settings}
\usage{
integrations_settings(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Retrieves a list of integration settings.
}
