% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\title{List}
\usage{
list(tok, fields = NULL, query = NULL)
}
\arguments{
\item{fields}{Field include hash (value of 1 to include, 0 to exclude). Optional}

\item{query}{Query filter hash. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets all of the users in the system and their information, the result is only limited to what the callee has access to view.. It supports the Offset, Count, and Sort Query Parameters along with Query and Fields Query Parameter.
}
