% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_list_departments.R
\name{list_departments}
\alias{list_departments}
\title{List departments}
\usage{
list_departments(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Get a list of departments.. It supports the Offset, Count, and Sort Query Parameters.
}
