% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_livechat_appearance_data.R
\name{livechat_appearance_data}
\alias{livechat_appearance_data}
\title{Livechat appearance data}
\usage{
livechat_appearance_data(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Get Livechat widget settings about appearance
}
