% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_dealing_with_livechat_agents.R
\name{livechat_dealing_with_livechat_agents}
\alias{livechat_dealing_with_livechat_agents}
\title{Dealing with Livechat agents}
\usage{
livechat_dealing_with_livechat_agents(tok, rid, token)
}
\arguments{
\item{rid}{The room _id. Required}

\item{token}{The visitor token. Required}
}
\description{
Retrieve the current Livechat agent data.
}
