% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_dealing_with_livechat_rooms.R
\name{livechat_dealing_with_livechat_rooms}
\alias{livechat_dealing_with_livechat_rooms}
\title{Dealing with Livechat rooms}
\usage{
livechat_dealing_with_livechat_rooms(tok, token, rid = NULL)
}
\arguments{
\item{token}{The visitor token. Required}

\item{rid}{The room _id. Optional}
}
\description{
Get the Livechat room data or open a new room.
}
