% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_dealing_with_livechat_transcript.R
\name{livechat_dealing_with_livechat_transcript}
\alias{livechat_dealing_with_livechat_transcript}
\title{Dealing with Livechat transcript}
\usage{
livechat_dealing_with_livechat_transcript(tok, token)
}
\arguments{
\item{token}{The visitor token Required}
}
\description{
Request a Livechat transcript
}
