% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_dealing_with_livechat_visitors.R
\name{livechat_dealing_with_livechat_visitors}
\alias{livechat_dealing_with_livechat_visitors}
\title{Dealing with Livechat visitors}
\usage{
livechat_dealing_with_livechat_visitors(tok, token)
}
\arguments{
\item{token}{The visitor token. Required}
}
\description{
Retrieve a visitor data
}
