% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_get_a_list_of_livechat_rooms.R
\name{livechat_get_a_list_of_livechat_rooms}
\alias{livechat_get_a_list_of_livechat_rooms}
\title{Get a list of livechat rooms}
\usage{
livechat_get_a_list_of_livechat_rooms(
  tok,
  agents = NULL,
  departmentid = NULL,
  open = NULL,
  createdat = NULL,
  closedat = NULL,
  tags = NULL,
  customfields = NULL,
  roomname = NULL
)
}
\arguments{
\item{agents}{A list of agents, can be a list of ids or usernames Optional}

\item{departmentid}{The department’s id Optional}

\item{open}{If it should filter by open/closed rooms Optional}

\item{createdat}{An object representing when the room was created(Can also be filtered only with start or end) Optional}

\item{closedat}{An object representing when the room was closed(Can also be filtered only with start or end) Optional}

\item{tags}{A list of tags Optional}

\item{customfields}{An object with custom fields to be filtered(previously created and populated at custom fields endpoints) Optional}

\item{roomname}{The room’s name Optional}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves a list of livechat rooms. It supports the Offset, Count, and Sort Query Parameters along with Fields Query Parameters.
}
