% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_inquiries_list.R
\name{livechat_inquiries_list}
\alias{livechat_inquiries_list}
\title{Inquiries List}
\usage{
livechat_inquiries_list(tok, department = NULL)
}
\arguments{
\item{department}{The department’s id or name Optional}

\item{token}{The token to connect to the app.}
}
\description{
Lists all of the open livechat inquiries. It supports the Offset, Count, and Sort Query Parameters.
}
