% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_list_agents_or_managers.R
\name{livechat_list_agents_or_managers}
\alias{livechat_list_agents_or_managers}
\title{List agents or managers}
\usage{
livechat_list_agents_or_managers(tok, type)
}
\arguments{
\item{type}{Can be either agent or manager. Required}

\item{token}{The token to connect to the app.}
}
\description{
Get a list of agents or managers. It supports the Offset, Count, and Sort Query Parameters.
}
