% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_livechat_configuration_data.R
\name{livechat_livechat_configuration_data}
\alias{livechat_livechat_configuration_data}
\title{Livechat configuration data}
\usage{
livechat_livechat_configuration_data(tok, token = NULL)
}
\arguments{
\item{token}{The visitor token. Optional}
}
\description{
Get Livechat widget configuration info and additional visitor data
}
