% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_livechat_office_hours.R
\name{livechat_office_hours}
\alias{livechat_office_hours}
\title{Livechat office hours}
\usage{
livechat_office_hours(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Get Livechat office hours
}
