% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_save_a_sms_message_on_rocket_chat.R
\name{livechat_save_a_sms_message_on_rocket_chat}
\alias{livechat_save_a_sms_message_on_rocket_chat}
\title{Save a SMS message on Rocket.Chat}
\usage{
livechat_save_a_sms_message_on_rocket_chat(tok, service)
}
\arguments{
\item{service}{Only the server twilio is accepted at moment. Required}

\item{token}{The token to connect to the app.}
}
\description{
Save a SMS message on Rocket.Chat
}
