% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechat_livechat_triggers.R
\name{livechat_triggers}
\alias{livechat_triggers}
\title{Livechat Triggers}
\usage{
livechat_triggers(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Lists all Livechat triggers. It supports the Offset, Count, and Sort Query Parameters.
}
