% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_directory.R
\name{miscellaneous_directory}
\alias{miscellaneous_directory}
\title{Directory}
\usage{
miscellaneous_directory(tok, query)
}
\arguments{
\item{query}{When type is users you can send an additional workspace field, that can be local (default) or all. workspace=all will work only if Federation is enabled. Required}

\item{token}{The token to connect to the app.}
}
\description{
A method, that searches by users or channels on all users and channels available on server.. It supports the Offset, Count, and Sort Query Parameters. along with Query and Fields Query Parameters.
}
