% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_shield_svg.R
\name{miscellaneous_shield_svg}
\alias{miscellaneous_shield_svg}
\title{Shield SVG}
\usage{
miscellaneous_shield_svg(
  tok,
  type = NULL,
  icon = NULL,
  channel = NULL,
  name = NULL
)
}
\arguments{
\item{type}{Can be one of online, user, channel Optional}

\item{icon}{Optional}

\item{channel}{Name of the channel Optional}

\item{name}{Name to show Optional}

\item{token}{The token to connect to the app.}
}
\description{
A simple method, requires no authentication, that returns the shield svg(badge) to add in your website. It must have API_Enable_Shields enabled.
}
