% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_spotlight.R
\name{miscellaneous_spotlight}
\alias{miscellaneous_spotlight}
\title{Spotlight}
\usage{
miscellaneous_spotlight(tok, query)
}
\arguments{
\item{query}{The term to be searched. Supports ‘#’ for channels and ‘@’ for users Required}

\item{token}{The token to connect to the app.}
}
\description{
Searches for users or rooms that are visible to the user.
}
