% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_statistics.R
\name{miscellaneous_statistics}
\alias{miscellaneous_statistics}
\title{Statistics}
\usage{
miscellaneous_statistics(tok, refresh = NULL)
}
\arguments{
\item{refresh}{Need to refresh statistics Optional}

\item{token}{The token to connect to the app.}
}
\description{
Statistics about the Rocket.Chat server. Requires view-statistics permission.
}
