% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_statistics_list.R
\name{miscellaneous_statistics_list}
\alias{miscellaneous_statistics_list}
\title{Statistics List}
\usage{
miscellaneous_statistics_list(
  tok,
  query,
  offset = NULL,
  count = NULL,
  fields = NULL,
  sort = NULL
)
}
\arguments{
\item{query}{See Query Parameter Required}

\item{offset}{See Offset Parameter Optional}

\item{count}{See Count Parameter Optional}

\item{fields}{See Fields Parameters Optional}

\item{sort}{See Sort Query Parameters Optional}

\item{token}{The token to connect to the app.}
}
\description{
Selectable statistics about the Rocket.Chat server. It supports the Offset, Count and Sort Query Parameters along with just the Fields and Query Parameters.. Requires view-statistics permission.
}
