% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauthapps_retrieves_an_oauth_app.R
\name{oauthapps_retrieves_an_oauth_app}
\alias{oauthapps_retrieves_an_oauth_app}
\title{Retrieves an OAuth App}
\usage{
oauthapps_retrieves_an_oauth_app(tok, appid, clientid)
}
\arguments{
\item{appid}{The app’s id. Required (if the clientId is not sent)}

\item{clientid}{The client’s id. Required (if the appId is not sent)}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves an OAuth App by id or client id
}
