% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions_update_permissions.R
\name{permissions_update_permissions}
\alias{permissions_update_permissions}
\title{Update Permissions}
\usage{
permissions_update_permissions(tok, permissions)
}
\arguments{
\item{permissions}{The array of roles to overwrite. Required}

\item{token}{The token to connect to the app.}
}
\description{
Edits permissions on the server.
}
