% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_private_settings_get.R
\name{private_settings_get}
\alias{private_settings_get}
\title{Private Settings Get}
\usage{
private_settings_get(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
List all private settings.
}
