% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_public_settings_get.R
\name{public_settings_get}
\alias{public_settings_get}
\title{Public Settings Get}
\usage{
public_settings_get(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
List all public settings.
}
