% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions_read.R
\name{read}
\alias{read}
\title{Read}
\usage{
read(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Mark messages as unread by roomId or from a message.. NOTE: You can send roomId or the id of the message (firstUnreadMessage._id) you want from it to be unread.
}
