% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenerate_personal_access_token.R
\name{regenerate_personal_access_token}
\alias{regenerate_personal_access_token}
\title{Regenerate Personal Access Token}
\usage{
regenerate_personal_access_token(tok, tokenname)
}
\arguments{
\item{tokenname}{The name of the token. Required}

\item{token}{The token to connect to the app.}
}
\description{
Regenerate a user personal access token. Requires create-personal-access-tokens permission.
}
