% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_avatar.R
\name{reset_avatar}
\alias{reset_avatar}
\title{Reset Avatar}
\usage{
reset_avatar(tok, userid_or_username)
}
\arguments{
\item{userid_or_username}{The id or username of the user. Required}

\item{token}{The token to connect to the app.}
}
\description{
Reset Avatar
}
