% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauthapps_retrieves_all_oauth_apps.R
\name{retrieves_all_oauth_apps}
\alias{retrieves_all_oauth_apps}
\title{Retrieves all OAuth Apps}
\usage{
retrieves_all_oauth_apps(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Retrieves all OAuth Apss
}
