% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_role_addusertorole.R
\name{roles_role_addusertorole}
\alias{roles_role_addusertorole}
\title{Role AddUserToRole}
\usage{
roles_role_addusertorole(tok, rolename, username, roomid = NULL)
}
\arguments{
\item{rolename}{The role name. Required}

\item{username}{The user name. Required}

\item{roomid}{If the role scope be Subscriptions and assign it to a room, you need to pass the roomId as parameter. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Assign a role to an user. Optionally, you can set this role to a room.
}
