% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_role_create.R
\name{roles_role_create}
\alias{roles_role_create}
\title{Role Create}
\usage{
roles_role_create(tok, name, scope = NULL, description = NULL)
}
\arguments{
\item{name}{The name of the new role. Required}

\item{scope}{The scope of the new role. Optional  Default: Users}

\item{description}{A description for the new role. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Create a new role in the system.
}
