% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_role_get_users_in_role.R
\name{roles_role_get_users_in_role}
\alias{roles_role_get_users_in_role}
\title{Role Get Users In Role}
\usage{
roles_role_get_users_in_role(tok, role, roomid = NULL)
}
\arguments{
\item{role}{The role. Required}

\item{roomid}{The room’s id. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets the users that belongs to a role. It supports the Offset and Count Only.
}
