% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_create_discussion.R
\name{rooms_create_discussion}
\alias{rooms_create_discussion}
\title{Rooms create discussion}
\usage{
rooms_create_discussion(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Creates a new discussion for room. It requires at least one of the following permissions: start-discussion OR start-discussion-other-user, AND must be with the following. setting enabled: Discussion_enabled.
}
