% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_favorite_unfavorite.R
\name{rooms_favorite_unfavorite}
\alias{rooms_favorite_unfavorite}
\title{Rooms Favorite/Unfavorite}
\usage{
rooms_favorite_unfavorite(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Favorite or unfavorite room.
}
