% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_get_discussions.R
\name{rooms_get_discussions}
\alias{rooms_get_discussions}
\title{Rooms Get Discussions}
\usage{
rooms_get_discussions(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Get all the discussions of room.  It supports the Offset, Count, and Sort Query Parameters.
}
