% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_clean_history.R
\name{rooms_rooms_clean_history}
\alias{rooms_rooms_clean_history}
\title{Rooms Clean History}
\usage{
rooms_rooms_clean_history(
  tok,
  roomid,
  latest,
  oldest,
  inclusive = NULL,
  excludepinned = NULL,
  filesonly = NULL,
  users = NULL
)
}
\arguments{
\item{roomid}{The room’s id Required}

\item{latest}{The end of time range of messages to clean Required}

\item{oldest}{The start of the time range of messages to clean Required}

\item{inclusive}{Whether messages which land on latest and oldest should be included Optional  Default: false}

\item{excludepinned}{Whether pinned messages should be deleted Optional  Default: false}

\item{filesonly}{Whether to only delete files and keep messages intact Optional  Default: false}

\item{users}{Specific set of users whose content to delete Optional  Default: [] (everyone)}

\item{token}{The token to connect to the app.}
}
\description{
Cleans up a room, removing messages from the provided time range.
}
