% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_info.R
\name{rooms_rooms_info}
\alias{rooms_rooms_info}
\title{Rooms Info}
\usage{
rooms_rooms_info(tok, roomid, roomname)
}
\arguments{
\item{roomid}{The rooms’s id Required (if no roomName)}

\item{roomname}{The rooms’s name Required (if no roomId)}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves the information about the room. It supports Fields Query Parameters only.
}
