% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_leave.R
\name{rooms_rooms_leave}
\alias{rooms_rooms_leave}
\title{Rooms Leave}
\usage{
rooms_rooms_leave(tok, roomid)
}
\arguments{
\item{roomid}{The room’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Causes the callee to be removed from the room.
}
