% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_save_notification_post.R
\name{rooms_rooms_save_notification_post}
\alias{rooms_rooms_save_notification_post}
\title{Rooms save notification POST}
\usage{
rooms_rooms_save_notification_post(
  tok,
  roomid,
  notifications_desktopnotifications = NULL,
  notifications_disablenotifications = NULL,
  notifications_emailnotifications = NULL,
  notifications_audionotifications = NULL,
  notifications_mobilepushnotifications = NULL,
  notifications_audionotificationvalue = NULL,
  notifications_desktopnotificationduration = NULL,
  notifications_unreadalert = NULL,
  notifications_hideunreadstatus = NULL
)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{notifications_desktopnotifications}{The value for desktop notifications Optional: ‘nothing’, ‘all’, ‘mentions’, ‘default’}

\item{notifications_disablenotifications}{The value to disable/enable channel’s notifications. Optional: ‘0’ or ‘1’}

\item{notifications_emailnotifications}{The value for email notifications. Optional: ‘nothing’, ‘all’, ‘mentions’, ‘default’}

\item{notifications_audionotifications}{The value for audio notifications. Optional: ‘nothing’, ‘all’, ‘mentions’, ‘default’}

\item{notifications_mobilepushnotifications}{The value for mobile push notifications. Optional: ‘nothing’, ‘all’, ‘mentions’, ‘default’}

\item{notifications_audionotificationvalue}{The value for audio notification sound. Optional}

\item{notifications_desktopnotificationduration}{The value for desktop notification duration. Optional}

\item{notifications_unreadalert}{The value for unread alerts notifications. Optional: ‘nothing’, ‘all’, ‘mentions’, ‘default’}

\item{notifications_hideunreadstatus}{The value to disable/enable number of unread messages. Optional: ‘0’ or ‘1’}

\item{token}{The token to connect to the app.}
}
\description{
Sets the notifications settings of specific channel.
}
