% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms_rooms_upload.R
\name{rooms_rooms_upload}
\alias{rooms_rooms_upload}
\title{Rooms Upload}
\usage{
rooms_rooms_upload(tok, file, msg = NULL, description = NULL, tmid = NULL)
}
\arguments{
\item{file}{A file name to upload Required}

\item{msg}{A message text Optional}

\item{description}{A description of the file Optional}

\item{tmid}{The thread message id (if you want upload a file to a thread) Optional}

\item{token}{The token to connect to the app.}
}
\description{
Post a message with attached file to a dedicated room.
}
