% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_service_configurations_get.R
\name{service_configurations_get}
\alias{service_configurations_get}
\title{Service Configurations Get}
\usage{
service_configurations_get(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
List all service configurations.
}
