% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_active_status.R
\name{set_active_status}
\alias{set_active_status}
\title{Set Active Status}
\usage{
set_active_status(tok, activestatus, userid)
}
\arguments{
\item{activestatus}{The value of the active status. Required}

\item{userid}{The user’s id to be changed. Required}

\item{token}{The token to connect to the app.}
}
\description{
Set Active Status
}
