% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_preferences.R
\name{set_preferences}
\alias{set_preferences}
\title{Set Preferences}
\usage{
set_preferences(tok, userid, data_newroomnotification,
  data_newmessagenotification, data_useemojis, data_convertasciiemoji,
  data_savemobilebandwidth, data_collapsemediabydefault,
  data_autoimageload, data_emailnotificationmode,
  data_roomslistexhibitionmode, data_unreadalert,
  data_notificationssoundvolume, data_desktopnotifications,
  data_mobilenotifications, data_enableautoaway, data_highlights,
  data_desktopnotificationduration,
  data_desktopnotificationrequireinteraction, data_viewmode,
  data_hideusernames, data_hideroles, data_hideavatars, data_sendonenter,
  data_roomcountersidebar, data_language,
  data_sidebarshowfavorites = NULL, data_sidebarshowunread = NULL,
  data_sidebarsortby = NULL, data_sidebarviewmode = NULL,
  data_sidebarhideavatar = NULL, data_groupbytype = NULL,
  data_mutefocusedconversations = NULL)
}
\arguments{
\item{userid}{The id of the user. Required}

\item{data_newroomnotification}{New room notification. Required}

\item{data_newmessagenotification}{New message notification. Required}

\item{data_useemojis}{User can use emojis. Required}

\item{data_convertasciiemoji}{Convert ascII emojis. Required}

\item{data_savemobilebandwidth}{Save mobile bandwidth. Required}

\item{data_collapsemediabydefault}{Collapse media by default. Required}

\item{data_autoimageload}{Image load automatically. Required}

\item{data_emailnotificationmode}{Email notification mode. Required}

\item{data_roomslistexhibitionmode}{Rooms list exhibition mode. Required}

\item{data_unreadalert}{Unread Alert. Required}

\item{data_notificationssoundvolume}{Volume of notification sound. Required}

\item{data_desktopnotifications}{Desktop notifications. Required}

\item{data_mobilenotifications}{Mobile notifications. Required}

\item{data_enableautoaway}{Enable auto away. Required}

\item{data_highlights}{Highlights. Required}

\item{data_desktopnotificationduration}{Duration of desktop notification. Required}

\item{data_desktopnotificationrequireinteraction}{Determines if user has to click on desktop notification to close it (requires Google Chrome version > 50 as client, overwrites setting desktopNotificationDuration). Required}

\item{data_viewmode}{View mode. Required}

\item{data_hideusernames}{Hide usernames. Required}

\item{data_hideroles}{Hide user roles. Required}

\item{data_hideavatars}{Hide avatars. Required}

\item{data_sendonenter}{Send message on enter. Required}

\item{data_roomcountersidebar}{Display room counter on sidebar. Required}

\item{data_language}{Language. Required}

\item{data_sidebarshowfavorites}{Show favorites on sidebar. Optional}

\item{data_sidebarshowunread}{Show unread on sidebar. Optional}

\item{data_sidebarsortby}{Show sort by. Optional}

\item{data_sidebarviewmode}{Show view mode. Optional}

\item{data_sidebarhideavatar}{Show avatar on hide bar. Optional}

\item{data_groupbytype}{Group channels by type. Optional}

\item{data_mutefocusedconversations}{Mute focused conversations. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Set Preferences
}
