% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_settings_oauth.R
\name{settings_oauth}
\alias{settings_oauth}
\title{Settings oAuth}
\usage{
settings_oauth(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
A simple method, requires no authentication, that returns list of all available oauth services.
}
