% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_settings_get.R
\name{settings_settings_get}
\alias{settings_settings_get}
\title{Settings Get}
\usage{
settings_settings_get(tok, x_id)
}
\arguments{
\item{x_id}{The setting’s _id field Required}

\item{token}{The token to connect to the app.}
}
\description{
Gets the setting for the provided _id.
}
