% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_settings_update.R
\name{settings_settings_update}
\alias{settings_settings_update}
\title{Settings Update}
\usage{
settings_settings_update(
  tok,
  x_id,
  value,
  color = NULL,
  editor,
  execute = NULL
)
}
\arguments{
\item{x_id}{The setting’s _id field Required}

\item{value}{The value to update the setting. Required}

\item{color}{The color code to update the setting. Optional}

\item{editor}{The editor key, with color value. Required if color was set}

\item{execute}{To trigger a action-button Optional}

\item{token}{The token to connect to the app.}
}
\description{
Updates the setting for the provided _id.
}
