% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions_subscriptions_get.R
\name{subscriptions_subscriptions_get}
\alias{subscriptions_subscriptions_get}
\title{Subscriptions Get}
\usage{
subscriptions_subscriptions_get(tok, updatedsince = NULL)
}
\arguments{
\item{updatedsince}{Date as ISO string Optional}

\item{token}{The token to connect to the app.}
}
\description{
Get all subscriptions. Please note that the unread counter value depends on your settings at the Administration => General section.
}
