% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions_subscriptions_get_one.R
\name{subscriptions_subscriptions_get_one}
\alias{subscriptions_subscriptions_get_one}
\title{Subscriptions Get one}
\usage{
subscriptions_subscriptions_get_one(tok, roomid)
}
\arguments{
\item{roomid}{The channel’s id Required}

\item{token}{The token to connect to the app.}
}
\description{
Get the subscription by room id.
}
