% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_own_basic_information.R
\name{update_own_basic_information}
\alias{update_own_basic_information}
\title{Update own basic information}
\usage{
update_own_basic_information(tok, data_email = NULL, data_name = NULL,
  data_username = NULL, data_currentpassword = NULL,
  data_newpassword = NULL, data_customfields = NULL)
}
\arguments{
\item{data_email}{The email address for the user. Optional}

\item{data_name}{The display name of the user. Optional}

\item{data_username}{The username for the user. Optional}

\item{data_currentpassword}{The password for the user encrypted in SHA256. Optional}

\item{data_newpassword}{The new password for the user Optional}

\item{data_customfields}{Any custom fields the user should have on their account. Optional  Default: undefined}

\item{token}{The token to connect to the app.}
}
\description{
Note that to update the password or email for the user you must send currentPassword property encrypted in SHA256 together in payload.
}
