% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_create_token.R
\name{users_create_token}
\alias{users_create_token}
\title{Create Token}
\usage{
users_create_token(tok, userid_or_username)
}
\arguments{
\item{userid_or_username}{The id or username of the user. Required}

\item{token}{The token to connect to the app.}
}
\description{
IMPORTANT. . To be able to use this endpoint you must set the ENV VAR (CREATE_TOKENS_FOR_USERS).
}
