% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_delete_own_account.R
\name{users_delete_own_account}
\alias{users_delete_own_account}
\title{Delete Own Account}
\usage{
users_delete_own_account(tok, password)
}
\arguments{
\item{password}{The password of user (encrypted in SHA256). Required}

\item{token}{The token to connect to the app.}
}
\description{
Deletes your own user. Requires Allow Users to Delete Own Account enabled. Accessible from Administration -> Accounts.
}
