% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_forgot_password.R
\name{users_forgot_password}
\alias{users_forgot_password}
\title{Forgot Password}
\usage{
users_forgot_password(tok, email)
}
\arguments{
\item{email}{The email to send password reset link Required}

\item{token}{The token to connect to the app.}
}
\description{
Send email to reset your password.
}
