% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_generate_personal_access_token.R
\name{users_generate_personal_access_token}
\alias{users_generate_personal_access_token}
\title{Generate Personal Access Token}
\usage{
users_generate_personal_access_token(tok, tokenname, bypasstwofactor = NULL)
}
\arguments{
\item{tokenname}{The name of the token. Required}

\item{bypasstwofactor}{If 2FA requirement should be ignored when using this token. Optional, default: false}

\item{token}{The token to connect to the app.}
}
\description{
Generate Personal Access Token. Requires create-personal-access-tokens permission.
}
