% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_get_avatar.R
\name{users_get_avatar}
\alias{users_get_avatar}
\title{Get Avatar}
\usage{
users_get_avatar(tok, userid_or_username)
}
\arguments{
\item{userid_or_username}{The id or username of the user. Required}

\item{token}{The token to connect to the app.}
}
\description{
Gets the URL for a user’s avatar.
}
