% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_get_presence.R
\name{users_get_presence}
\alias{users_get_presence}
\title{Get Presence}
\usage{
users_get_presence(tok, userid_or_username = NULL)
}
\arguments{
\item{userid_or_username}{The id or username of the user. If not provided, the auth user is updated. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets a user’s presence if the query string userId or username is provided, otherwise it gets the callee’s.
}
