% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_presence.R
\name{users_presence}
\alias{users_presence}
\title{Users Presence}
\usage{
users_presence(tok, from = NULL)
}
\arguments{
\item{from}{The last date you got a status change Optional}

\item{token}{The token to connect to the app.}
}
\description{
Gets all connected users presence.
}
