% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_register.R
\name{users_register}
\alias{users_register}
\title{Register}
\usage{
users_register(tok, username, email, pass, name, secreturl = NULL)
}
\arguments{
\item{username}{The username for the user. Required}

\item{email}{The email for the user. Required}

\item{pass}{The password for the user. Required}

\item{name}{The name of the user. Required}

\item{secreturl}{String appended to secret registration URL (if using). Optional}

\item{token}{The token to connect to the app.}
}
\description{
Register
}
