% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_set_avatar.R
\name{users_set_avatar}
\alias{users_set_avatar}
\title{Set Avatar}
\usage{
users_set_avatar(tok, image, userid_or_username = NULL)
}
\arguments{
\item{image}{The image file to use as the new avatar, as form data. Required}

\item{userid_or_username}{The id or username of the user. If not provided, the auth user is updated. Optional}

\item{token}{The token to connect to the app.}
}
\description{
Set Avatar
}
