% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_conference_update_jitsi_timeout.R
\name{video_conference_update_jitsi_timeout}
\alias{video_conference_update_jitsi_timeout}
\title{Update Jitsi Timeout}
\usage{
video_conference_update_jitsi_timeout(tok, roomid)
}
\arguments{
\item{roomid}{The room Id of the room to update the jitsi timeout. Required}

\item{token}{The token to connect to the app.}
}
\description{
Update Jitsi Timeout
}
