% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav_webdav_get_my_accounts.R
\name{webdav_get_my_accounts}
\alias{webdav_get_my_accounts}
\title{Webdav Get my Accounts}
\usage{
webdav_get_my_accounts(tok)
}
\arguments{
\item{token}{The token to connect to the app.}
}
\description{
Retrieves the user’s webdav accounts.
}
