% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Info}
\usage{
info(tok, userid_or_username)
}
\arguments{
\item{userid_or_username}{The id or username of the user. Required}

\item{token}{The token to connect to the app.}
}
\description{
Retrieves information about a user, the result is only limited to what the callee has access to view.. It supports  Fields Query Parameter with the userRooms field, that returns the rooms that the user is part of.
}
